Refactor isValidSudoku solution
Rows hashmap was removed because a set can be created for each row
while iterating to keep track of duplicates in each row. There is no
need to persist data of previous rows - if it has a duplicate the
function will short-circuit, if it doesn't then there is no further
processing to be done.

Values can be assigned to grids (subgrids) using a clever integer
division trick to avoid using a hard-coded helper function. The
row and column of each value is divided by the number of rows/cols
of subgrids in the board, and the integer remainders of each ranges
from [0, (number of row/cols) - 1]. This gives a tuple (row, col)
defining which subgrid on the board the value lies in.