Refactor - renaming and removing current_num var.
Removed the variable current_num that checked the current number
in the sequence we are at (check if it is present in the number
set or not) because the next number in a sequence that increases
by 1 can be calculated by adding the starting number of the
sequence with its current length - which we are already keeping
track of.

In addition, renamed max_len_seq to longest and current_len_seq to
current_length for readability.