Add solution to min stack coding problem.
The challenge is to implement a getMin() function, that gets the minimum
value in the stack, in O(1) time. The trick is to associate each node
(val) with the current min value. In other words, keep track of the
current min with each insertion. This can be done by implementing a
second stack that is 1:1 with the original stack but it only stores the
min value at each position/insertion in the stack. Thus when you pop
from the main stack, you concurrently pop from the min stack and have
the min value in the stack at the time the value at the top of the stack
was inserted.