Remove unnecessary logic and naming changes.
Removed conditional in nested while-loop that checked if the previous
height (i.e. height on top of stack) was greater than the current
height. This check is already performed in the while-loop condition, so
the conditional is unnecessary.

Variable assignments in the beginning of while-loop are combined into
the stack.pop() statement atthe end of the loop. If an item is popped
from the stack regardless at the end of each while-loop iteration,
then it is more concise and readable to assign the variables by
unpacking them in the stack.pop() call (can combine into a single
line).

Variable names are abbreviated, and index,height pairs in stack are
represented as tuples instead of nested lists (less memory, slightly
faster).

The fundamental trick for this problem is figuring out how to get the
left boundary for each height (as getting max area only moving forward
(not looking at previous heights)) is relatively simple - once you hit a
height that is lower than the current, pop from stack as long as this
condition holds and compute the areas. The trick is to reassign the
index of the current height to each height you pop from the stack as long
as they are greater than the current height (i.e. the current height
'stretches' back until you reach a height in the stack that is lower
than the current height).